Refactor all usages of API hooks to be used from folder
