+++ src/api/hbvResearch/index.ts
@@ -1,0 +1,3 @@
+import useHbvResearchPapers from './useHbvResearchPapers';
+
+export { useHbvResearchPapers };
