+++ src/api/holisticOffice/index.ts
@@ -1,0 +1,4 @@
+import useHolisticOfficeLinks from './useHolisticOfficeLinks';
+import useHolisticOfficeModules from './useHolisticOfficeModules';
+
+export { useHolisticOfficeLinks, useHolisticOfficeModules };
