+++ src/api/martialArts/index.ts
@@ -1,0 +1,3 @@
+import useMartialArtsStyles from './useMartialArtsStyles';
+
+export { useMartialArtsStyles };
