+++ src/api/music/index.ts
@@ -1,0 +1,4 @@
+import useInstruments from './useInstruments';
+import useMusicScores from './useMusicScores';
+
+export { useInstruments, useMusicScores };
