+++ src/components/AppContent/HolisticOfficePage/ModulesGrid.tsx
@@ -1,5 +1,5 @@
 import { Grid } from '@material-ui/core';
-import useHolisticOfficeModules from 'api/holisticOffice/useHolisticOfficeModules';
+import { useHolisticOfficeModules } from 'api/holisticOffice';
 import ErrorView from 'components/shared/ErrorView';
 import LoadingView from 'components/shared/LoadingView';
 import { isNil } from 'lodash';
