+++ src/components/AppContent/MusicScoresPage/SectionGrid/SectionGrid.tsx
@@ -1,6 +1,6 @@
 import { Zoom } from '@material-ui/core';
 import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
-import useInstruments from 'api/music/useInstruments';
+import { useInstruments } from 'api/music';
 import ErrorView from 'components/shared/ErrorView';
 import { ResponsiveGrid, ResponsiveGridItem } from 'components/shared/ResponsiveGrid';
 import { isNil, map } from 'lodash';
